+++ src/core/jit/transform/check_batch_loader.rs
