+++ src/core/jit/transform/mod.rs
@@ -1,7 +1,7 @@
 mod check_cache;
 mod check_const;
 mod check_dedupe;
-mod check_direct_loader;
+mod check_batch_loader;
 mod check_protected;
 mod input_resolver;
 mod skip;
@@ -9,7 +9,7 @@ mod skip;
 pub use check_cache::*;
 pub use check_const::*;
 pub use check_dedupe::*;
-pub use check_direct_loader::*;
+pub use check_batch_loader::*;
 pub use check_protected::*;
 pub use input_resolver::*;
 pub use skip::*;
